adds comments
