+++ templates/frontend/gateway/src/App.js
@@ -1,8 +1,7 @@
 import "App.scss";
 import "bootstrap-icons/font/bootstrap-icons.css";
 import "bootstrap/dist/js/bootstrap.bundle";
-import { Widget } from "near-social-vm";
-import React, { useEffect, useMemo } from "react";
+import React, { useEffect } from "react";
 import "react-bootstrap-typeahead/css/Typeahead.css";
 
 import { sanitizeUrl } from "@braintree/sanitize-url";
@@ -10,47 +9,16 @@ import { useAccount, useInitNear } from "near-social-vm";
 import {
   createBrowserRouter,
   Link,
-  RouterProvider,
-  useLocation,
+  Navigate,
+  RouterProvider
 } from "react-router-dom";
-import useRedirectMap from "./hooks/useRedirectMap";
-import useUrbitShip from "./hooks/useUrbitShip";
-
-function Viewer({ widgetSrc, code, initialProps }) {
-  const location = useLocation();
-  const searchParams = new URLSearchParams(location.search);
-  const { components: redirectMap } = useRedirectMap();
-  const api = useUrbitShip();
-
-  // create props from params
-  const passProps = useMemo(() => {
-    return Array.from(searchParams.entries()).reduce((props, [key, value]) => {
-      props[key] = value;
-      return props;
-    }, {});
-  }, [location]);
-
-  const path = location.pathname.substring(1);
-
-  const src = useMemo(() => {
-    const pathSrc = widgetSrc ?? path;
-    return pathSrc;
-  }, [widgetSrc, path]);
-
-  return (
-    <>
-      <Widget
-        src={!code && src}
-        code={code} // prioritize code
-        props={{ ...initialProps, ...passProps, api }}
-        config={{ redirectMap }}
-      />
-    </>
-  );
-}
 
 function App(props) {
+  // Attributes passed from web component
   const { src, code, initialProps, rpc, network, selectorPromise } = props;
+
+  // Initialize the NEAR connection
+  // and the VM itself
   const { initNear } = useInitNear();
 
   useAccount();
@@ -71,7 +39,7 @@ function App(props) {
         },
       },
       features: {
-        enableComponentSrcDataKey: true,
+        enableComponentSrcDataKey: true, // adds "data-component=${src}" attribute to widgets, helpful for inspect element
       },
       config: {
         defaultFinality: undefined,
@@ -79,27 +47,29 @@ function App(props) {
     };
 
     if (rpc) {
-      config.config.nodeUrl = rpc;
+      config.config.nodeUrl = rpc; // <Widget src="..." /> makes an rpc request for the widget code
+      // this enables us to override this rpc url to use a local, rpc proxy (bos-workspace)
     }
 
     initNear && initNear(config);
   }, [initNear, rpc]);
 
+  // Specific to the Urbit implementation
   let pathname = window.location.pathname;
   let before = pathname.substring(0, pathname.indexOf(`/gateway`));
 
   const router = createBrowserRouter([
     {
-      path: "/*",
+      path: "/",
+      element: <Navigate to="/placeholder/gateway/" />,
+    },
+    {
+      path: `/${before || "placeholder"}/gateway/*`,
       element: (
         <Viewer widgetSrc={src} code={code} initialProps={initialProps} />
       ),
     },
-  ],
-  {
-    basename: `${before}/gateway/` // Why does this need to be here?
-  }
-);
+  ]);
 
   return <RouterProvider router={router} />;
 }
